Revert "Consolidate library files into `build_cli.dart`"
This reverts commit 81748c032d77adabed1dc3cb345bc695fcaee5f1.

Hold to the convention – builder.dart has the builder
