+++ build_cli/CHANGELOG.md
@@ -1,12 +1,9 @@
-## 0.2.0
+## 0.1.1
 
 - Add setup instructions to `README.md`.
 - Use the logger built into `pkg:build`.
 - Don't generate CLI entries for unassignable fields.
 - Support latest `pkg:source_gen`. 
-- *BREAKING* (strictly speaking, but not likely.)
-  - Removed `package:build_cli/builder.dart`.
-  - Moved and renamed the builder factory to `build_cli.dart`.
 
 ## 0.1.0
