+++ build_cli/lib/build_cli.dart
@@ -7,39 +7,4 @@
 /// for more information.
 library build_cli;
 
-import 'package:build/build.dart';
-import 'package:logging/logging.dart';
-import 'package:source_gen/source_gen.dart';
-
-import 'src/build_cli_generator.dart';
-
 export 'src/build_cli_generator.dart' show CliGenerator;
-
-// TODO: until we can use `log` here - github.com/dart-lang/build/issues/1223
-final _logger = new Logger('build_cli');
-
-/// Supports `package:build_runner` creation and configuration of `build_cli`.
-///
-/// Not meant to be invoked by hand-authored code.
-Builder createBuilderForCliGenerator(BuilderOptions options) {
-  // Paranoid copy of options.config - don't assume it's mutable or needed
-  // elsewhere.
-  var optionsMap = new Map<String, dynamic>.from(options.config);
-
-  try {
-    return _cliPartBuilder(header: optionsMap.remove('header') as String);
-  } finally {
-    if (optionsMap.isNotEmpty) {
-      _logger.warning('These options were ignored: `$optionsMap`.');
-    }
-  }
-}
-
-Builder _cliPartBuilder({String header}) {
-  return new PartBuilder(
-    const [
-      const CliGenerator(),
-    ],
-    header: header,
-  );
-}
